+++ packages/schema/src/query_response.rs
@@ -1,9 +1,6 @@
 use std::collections::{BTreeMap, BTreeSet};
 
-use schemars::{
-    schema::{InstanceType, RootSchema, SingleOrVec, SubschemaValidation},
-    JsonSchema,
-};
+use schemars::{schema::RootSchema, JsonSchema};
 use thiserror::Error;
 
 pub use cosmwasm_schema_derive::QueryResponses;
@@ -69,10 +66,6 @@ pub trait QueryResponses: JsonSchema {
     fn response_schemas() -> Result<BTreeMap<String, RootSchema>, IntegrityError> {
         let response_schemas = Self::response_schemas_impl();
 
-        let queries: BTreeSet<_> = response_schemas.keys().cloned().collect();
-
-        check_api_integrity::<Self>(queries)?;
-
         Ok(response_schemas)
     }
 
@@ -93,121 +86,6 @@ pub fn combine_subqueries<const N: usize, T>(
         )
     }
     map
-}
-
-/// Returns possible enum variants from `one_of` analysis
-fn enum_variants(
-    subschemas: SubschemaValidation,
-) -> Result<impl Iterator<Item = Result<String, IntegrityError>>, IntegrityError> {
-    let iter = subschemas
-        .one_of
-        .ok_or(IntegrityError::InvalidQueryMsgSchema)?
-        .into_iter()
-        .map(|s| {
-            let s = s.into_object();
-
-            if let Some(SingleOrVec::Single(ty)) = s.instance_type {
-                match *ty {
-                    // We'll have an object if the Rust enum variant was C-like or tuple-like
-                    InstanceType::Object => s
-                        .object
-                        .ok_or(IntegrityError::InvalidQueryMsgSchema)?
-                        .required
-                        .into_iter()
-                        .next()
-                        .ok_or(IntegrityError::InvalidQueryMsgSchema),
-                    // We might have a string here if the Rust enum variant was unit-like
-                    InstanceType::String => {
-                        let values = s.enum_values.ok_or(IntegrityError::InvalidQueryMsgSchema)?;
-
-                        if values.len() != 1 {
-                            return Err(IntegrityError::InvalidQueryMsgSchema);
-                        }
-
-                        values[0]
-                            .as_str()
-                            .map(String::from)
-                            .ok_or(IntegrityError::InvalidQueryMsgSchema)
-                    }
-                    _ => Err(IntegrityError::InvalidQueryMsgSchema),
-                }
-            } else {
-                Err(IntegrityError::InvalidQueryMsgSchema)
-            }
-        });
-
-    Ok(iter)
-}
-
-fn verify_queries(
-    query_msg: BTreeSet<String>,
-    responses: BTreeSet<String>,
-) -> Result<(), IntegrityError> {
-    if query_msg != responses {
-        return Err(IntegrityError::InconsistentQueries {
-            query_msg,
-            responses,
-        });
-    }
-
-    Ok(())
-}
-
-/// `generated_queries` is expected to be a sorted slice here!
-fn check_api_integrity<T: QueryResponses + ?Sized>(
-    generated_queries: BTreeSet<String>,
-) -> Result<(), IntegrityError> {
-    let schema = crate::schema_for!(T);
-
-    let subschemas = if let Some(subschemas) = schema.schema.subschemas {
-        subschemas
-    } else {
-        // No subschemas - no resposnes are expected
-        return verify_queries(BTreeSet::new(), generated_queries);
-    };
-
-    let schema_queries = if let Some(any_of) = subschemas.any_of {
-        // If `any_of` exists, we assume schema is generated from untagged enum
-        any_of
-            .into_iter()
-            .map(|schema| dbg!(schema.into_object()))
-            .filter_map(|obj| {
-                if let Some(reference) = obj.reference {
-                    // Subschemas can be hidden behind references - we want to map them to proper
-                    // subschemas in such case
-
-                    // Only references to definitions are supported
-                    let reference = match reference.strip_prefix("#/definitions/") {
-                        Some(reference) => reference,
-                        None => {
-                            return Some(Err(IntegrityError::ExternalReference {
-                                reference: reference.to_owned(),
-                            }))
-                        }
-                    };
-
-                    let schema = match schema.definitions.get(reference) {
-                        Some(schema) => schema.clone(),
-                        None => return Some(Err(IntegrityError::InvalidQueryMsgSchema)),
-                    };
-
-                    Ok(schema.into_object().subschemas).transpose()
-                } else {
-                    Ok(obj.subschemas).transpose()
-                }
-            })
-            .map(|subschema| enum_variants(*subschema?))
-            .collect::<Result<Vec<_>, _>>()?
-            .into_iter()
-            .flatten()
-            .collect::<Result<_, _>>()?
-    } else {
-        // If `any_of` is not present, there was no untagged enum on top, we expect normal enum at
-        // this point
-        enum_variants(*subschemas)?.collect::<Result<_, _>>()?
-    };
-
-    verify_queries(schema_queries, generated_queries)
 }
 
 #[derive(Debug, Error, PartialEq, Eq)]
@@ -297,18 +175,6 @@ mod tests {
         fn response_schemas_impl() -> BTreeMap<String, RootSchema> {
             BTreeMap::from([("balance_for".to_string(), schema_for!(u128))])
         }
-    }
-
-    #[test]
-    fn bad_msg_fails() {
-        let err = BadMsg::response_schemas().unwrap_err();
-        assert_eq!(
-            err,
-            IntegrityError::InconsistentQueries {
-                query_msg: BTreeSet::from(["balance-for".to_string()]),
-                responses: BTreeSet::from(["balance_for".to_string()])
-            }
-        );
     }
 
     #[derive(Debug, JsonSchema)]