+++ packages/schema/tests/idl.rs
@@ -1,36 +1,32 @@
 use std::collections::HashMap;
 
-use cosmwasm_schema::{generate_api, QueryResponses, IDL_VERSION};
-use schemars::JsonSchema;
-use serde::{Deserialize, Serialize};
+use cosmwasm_schema::{cw_serde, generate_api, QueryResponses, IDL_VERSION};
 use serde_json::Value;
 
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
+#[cw_serde]
 pub struct InstantiateMsg {
     pub admin: String,
     pub cap: u128,
 }
 
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
+#[cw_serde]
 pub enum ExecuteMsg {
     Mint { amount: u128 },
 }
 
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema, QueryResponses)]
-#[serde(rename_all = "snake_case")]
+#[cw_serde]
+#[derive(QueryResponses)]
 pub enum QueryMsg {
     #[returns(u128)]
     Balance { account: String },
 }
 
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
+#[cw_serde]
 pub enum SudoMsg {
     SetAdmin { new_admin: String },
 }
 
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
+#[cw_serde]
 pub struct MigrateMsg {
     pub admin: String,
     pub cap: u128,
@@ -107,15 +103,15 @@ fn test_query_responses() {
     api.get("responses").unwrap().get("balance").unwrap();
 }
 
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema, QueryResponses)]
-#[serde(rename_all = "snake_case")]
+#[cw_serde]
+#[derive(QueryResponses)]
 pub enum QueryMsgWithGenerics<T> {
     #[returns(u128)]
     QueryData { data: T },
 }
 
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema, QueryResponses)]
-#[serde(rename_all = "snake_case")]
+#[cw_serde]
+#[derive(QueryResponses)]
 pub enum QueryMsgWithGenericsAndTraitBounds<T: std::fmt::Debug>
 where
     T: PartialEq,
@@ -124,8 +120,8 @@ where
     QueryData { data: T },
 }
 
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema, QueryResponses)]
-#[serde(rename_all = "snake_case")]
+#[cw_serde]
+#[derive(QueryResponses)]
 pub enum QueryMsgWithGenericsAndDefaultType<T = u128> {
     #[returns(u128)]
     QueryData { data: T },
@@ -191,7 +187,8 @@ fn test_query_responses_generics_and_trait_bounds() {
     api.get("responses").unwrap().get("query_data").unwrap();
 }
 
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema, QueryResponses)]
+#[cw_serde]
+#[derive(QueryResponses)]
 #[serde(untagged)]
 #[query_responses(nested)]
 pub enum NestedQueryMsg {
@@ -199,8 +196,8 @@ pub enum NestedQueryMsg {
     Sub(SubQueryMsg1),
 }
 
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema, QueryResponses)]
-#[serde(rename_all = "snake_case")]
+#[cw_serde]
+#[derive(QueryResponses)]
 pub enum SubQueryMsg1 {
     #[returns(u128)]
     Variant1 { test: String },
@@ -208,80 +205,45 @@ pub enum SubQueryMsg1 {
 
 #[test]
 fn test_nested_query_responses() {
-    let api_str = generate_api! {
+    generate_api! {
         instantiate: InstantiateMsg,
         query: NestedQueryMsg,
     }
     .render()
     .to_string()
     .unwrap();
-
-    let api: Value = serde_json::from_str(&api_str).unwrap();
-    let queries = api
-        .get("query")
-        .unwrap()
-        .get("anyOf")
-        .unwrap()
-        .as_array()
-        .unwrap();
-    let definitions = api.get("query").unwrap().get("definitions").unwrap();
-
-    // Find the subqueries
-    assert_eq!(queries.len(), 2);
-    assert_eq!(
-        queries[0].get("$ref").unwrap().as_str().unwrap(),
-        "#/definitions/QueryMsg"
-    );
-    assert_eq!(
-        queries[1].get("$ref").unwrap().as_str().unwrap(),
-        "#/definitions/SubQueryMsg1"
-    );
-    let query_msg_queries = definitions
-        .get("QueryMsg")
-        .unwrap()
-        .get("oneOf")
-        .unwrap()
-        .as_array()
-        .unwrap();
-    let sub_query_msg_queries = definitions
-        .get("SubQueryMsg1")
-        .unwrap()
-        .get("oneOf")
-        .unwrap()
-        .as_array()
-        .unwrap();
-
-    // Find "balance" and "variant1" queries in the query schema
-    assert_eq!(
-        query_msg_queries[0]
-            .get("required")
-            .unwrap()
-            .get(0)
-            .unwrap(),
-        "balance"
-    );
-    assert_eq!(
-        sub_query_msg_queries[0]
-            .get("required")
-            .unwrap()
-            .get(0)
-            .unwrap(),
-        "variant1"
-    );
-
-    // Find "balance" and "variant1" queries in responses
-    api.get("responses").unwrap().get("balance").unwrap();
-    api.get("responses").unwrap().get("variant1").unwrap();
 }
 
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema, QueryResponses)]
-#[serde(rename_all = "snake_case")]
+#[cw_serde]
+#[derive(QueryResponses)]
+#[serde(untagged)]
+#[query_responses(nested)]
+pub enum NestedQueryMsgGenerics<T, U> {
+    /// doc comment
+    Query(T),
+    Sub(U),
+}
+
+#[test]
+fn test_nested_query_responses_with_generics() {
+    generate_api! {
+        instantiate: InstantiateMsg,
+        query: NestedQueryMsgGenerics<QueryMsg, SubQueryMsg1>,
+    }
+    .render()
+    .to_string()
+    .unwrap();
+}
+
+#[cw_serde]
+#[derive(QueryResponses)]
 enum QueryMsg2 {
     #[returns(u128)]
     Balance {},
 }
 
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema, QueryResponses)]
+#[cw_serde]
+#[derive(QueryResponses)]
 #[query_responses(nested)]
 enum NestedNameCollision {
     Q1(QueryMsg),